split new draw code into multiple files
